Move WindowLocaleEffectManager to stepper
