+++ client/landing/stepper/utils/window-locale-effect-manager.tsx
@@ -1,7 +1,7 @@
 import { useLocale } from '@automattic/i18n-utils';
 import { useI18n } from '@wordpress/react-i18n';
 import * as React from 'react';
-import { switchWebpackCSS } from '../../../../lib/i18n-utils/switch-locale';
+import { switchWebpackCSS } from '../../../lib/i18n-utils/switch-locale';
 
 export const WindowLocaleEffectManager: React.FunctionComponent = () => {
 	const { __, _x, isRTL } = useI18n();
